+++ auth/email_action_links.go
@@ -1,0 +1,122 @@
+// Copyright 2019 Google Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package auth
+
+import (
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"net/http"
+	"net/url"
+)
+
+// ActionCodeSettings specifies the required continue/state URL with optional Android and iOS settings. Used when
+// invoking the email action link generation APIs.
+type ActionCodeSettings struct {
+	URL                   string `json:"continueUrl"`
+	HandleCodeInApp       bool   `json:"canHandleCodeInApp"`
+	IOSBundleID           string `json:"iOSBundleId,omitempty"`
+	AndroidPackageName    string `json:"androidPackageName,omitempty"`
+	AndroidMinimumVersion string `json:"androidMinimumVersion,omitempty"`
+	AndroidInstallApp     bool   `json:"androidInstallApp,omitempty"`
+	DynamicLinkDomain     string `json:"dynamicLinkDomain,omitempty"`
+}
+
+func (settings *ActionCodeSettings) toMap() (map[string]interface{}, error) {
+	if settings.URL == "" {
+		return nil, errors.New("URL must not be empty")
+	}
+
+	url, err := url.Parse(settings.URL)
+	if err != nil || url.Scheme == "" || url.Host == "" {
+		return nil, fmt.Errorf("malformed url string: %q", settings.URL)
+	}
+
+	if settings.AndroidMinimumVersion != "" || settings.AndroidInstallApp {
+		if settings.AndroidPackageName == "" {
+			return nil, errors.New("Android package name is required when specifying other Android settings")
+		}
+	}
+
+	b, err := json.Marshal(settings)
+	if err != nil {
+		return nil, err
+	}
+	var result map[string]interface{}
+	if err := json.Unmarshal(b, &result); err != nil {
+		return nil, err
+	}
+	return result, nil
+}
+
+type linkType string
+
+const (
+	emailLinkSignIn   linkType = "EMAIL_SIGNIN"
+	emailVerification linkType = "VERIFY_EMAIL"
+	passwordReset     linkType = "PASSWORD_RESET"
+)
+
+// EmailVerificationLink generates the out-of-band email action link for email verification flows for the specified
+// email address.
+func (c *userManagementClient) EmailVerificationLink(ctx context.Context, email string) (string, error) {
+	return c.EmailVerificationLinkWithSettings(ctx, email, nil)
+}
+
+// EmailVerificationLinkWithSettings generates the out-of-band email action link for email verification flows for the
+// specified email address, using the action code settings provided.
+func (c *userManagementClient) EmailVerificationLinkWithSettings(
+	ctx context.Context, email string, settings *ActionCodeSettings) (string, error) {
+	return c.generateEmailActionLink(ctx, emailVerification, email, settings)
+}
+
+func (c *userManagementClient) generateEmailActionLink(
+	ctx context.Context, linkType linkType, email string, settings *ActionCodeSettings) (string, error) {
+
+	if email == "" {
+		return "", errors.New("email must not be empty")
+	}
+
+	payload := map[string]interface{}{
+		"requestType":   linkType,
+		"email":         email,
+		"returnOobLink": true,
+	}
+	if settings != nil {
+		settingsMap, err := settings.toMap()
+		if err != nil {
+			return "", err
+		}
+		for k, v := range settingsMap {
+			payload[k] = v
+		}
+	}
+
+	resp, err := c.post(ctx, "/accounts:sendOobCode", payload)
+	if err != nil {
+		return "", err
+	}
+
+	if resp.Status != http.StatusOK {
+		return "", handleHTTPError(resp)
+	}
+
+	var result struct {
+		OOBLink string `json:"oobLink"`
+	}
+	err = json.Unmarshal(resp.Body, &result)
+	return result.OOBLink, err
+}